main_project/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx~=0.26.0
pydantic>=2.0.0
python-dotenv>=1.0.1
apscheduler==3.10.4
python-telegram-bot==20.8
motor==3.3.2
pymongo==4.6.1
rx>=3.2.0
requests>=2.31.0
pytz

# LangChain dependencies for LLM integration
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.5
# DeepSeek integration (using OpenAI-compatible API)
langchain-core>=0.1.0
# Additional LangChain tools and utilities
langchain-experimental>=0.0.1

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
# Add any other dependencies as needed 